Add script to un-rewrite part of the deps
Specifically written for "opentracing", because it is used in some
libp2p project and so every
"import github.com/opentracing/opentracing-go" is rewritten in our
project. However, we use `appdash` which is not used in libp2p projects,
and is not rewritten when `gx-go rw`. This introduce conflict of types.
So a temporary solution is to unwrite "opentracin-go" after `gx-go rw`